Change the spec of callbacks: msg-type is str now
Test ein:notebook-execute-current-cell is passed.
At this point all unit tests in test/test-ein-*.el are passed.

msg-type was first a string and I changed it to a keyword when I am
writing ein:kernel-execute.  But as I need to fix a lot of other
places, I change it back to a string.  Note that msg-type can come
from JSON directly when the notebook is first opened, so it must be
consistent with that routines.

Maybe I will change it again, but at that time, symbol is better than
keyword because msg-type is a value rather than a key.