ein:notebooklist-delete-notebook works again
The bug was in ein:query-ajax-callback.  When url-retrieve called in
its callback (which is the case of ein:notebooklist-delete-notebook),
same process buffer is reused.  Therefore, old code killed another
buffer (namely notebook list buffer) by accident and the variable
url-http-response-status was void because it is outside of the process
buffer (I am not sure about this part).  Current workaround is to
check if the buffer is live before executing buffer, using new macro
ein:with-live-buffer.