Add ein:pytools-jump-to-source-command (M-.)
As this command requires to invoke multi-line python code, ein.py is
added to store python code needed for EIN.  The sys.path is added
automatically when the kernel is started so that any function can be
called as ``__import__("ein").FUNCTION()``.  This way, namespace is
not contaminated at all.