Use hash table for ein:notebooklist-list
In the previous version, the global variable ein:notebooklist-list
kept growing when the notebook list is reloaded.  In this change
backend for data storage is switched to hash table so that the size
of data will not increase when the list is reloaded.

Fixes the problem old notebook name showed up in the completion
list of ein:notebooklist-open-notebook-global.