Use overlay to specify ein:cell-heading-N faces
As there is some problem with text property based face and MuMaMo,
using text property for heading is not good idea.  Turing on/off
heading face changes height and moves the following contents in the
notebook.  This is not allowable.

Furthermore, previous approach does not work when there is no text
in the input area or when inserting in front of the current heading.