Fix the problem with autoexec and insert
The problem was that ein:cell-input-pos-min/max fail when called
during the insertion, because the input area does not exist yet
when inserting, for example prompt EWOC node.  This is fixed in
the previous commit 0205188933bf0950b95feba367126ecbec2f6deb.
They return nil when there is no input area yet.  So,
ein:autoexec-after-change checks if they return nil or not now.