Include scratch sheet in ein:notebook-buffer
Now it is possible to have worksheet buffer closed but scratchsheet
buffer opened.  In this case, ein:notebook-live-p should return t
and ein:notebook-buffer should return the scratchsheet buffer.
This is because notebook commands can such as save or rename can be
executed from the scratchsheet buffer.

ein:notebook-to-json-after-closing-a-worksheet is modified also
this test fails without the change in ein:notebook-buffer.