Fix ein:testing-delete-notebook-by-name
ein:notebooklist-reload, which is called from success
callback of ein:notebooklist-delete-notebook, needs to
access ein:%notebooklist%.  However it was set to nil
in ein:testing-delete-notebook-by-name just after the
delete button is "clicked".  This was meant to detect
if the delete process is finished, but was not working.
To solve this problem, I add ein:notebooklist-after-open-hook
and use it to detect end of the delete process.