Bug fix, better info on running kernel.
Fixed bug with `ein:notebook-open-in-browser`; this command should now
work in IPython notebook 3.0 and later.

Notebook/worksheets now display the name of the kernel that the notebook
is running under. Next step is to make this clickable!

Also some documentation for my new super-secret project to add
evaluation of dynamic javascript to ein.