Basic javascript support via skewer.el
EIN will try to use skewer to execute javascript cells, which works for
the most part. Those hoping to have widgest or something like d3 working
are going to be disappointed, however, as at its heart ein is not an
HTML page and those packages depend on modifying the DOM of the notebook
web page.

It remains to be seen if those issues are surmountable, but for the
moment this functionality will live in this branch and this branch
alone.