Introduce skewer-mode dependency
I swear I though I was on another branch when I commited the dynamic javascript
stuff but, well, nothing to be done now I suppose.

With very basic support for evaluating javascript code now available I am adding
the skewer-mode dependcy as I should have done originally.

Note that yes, one can execute javascript in code blocks with the %%javascript
magic, but don't expect the really cools things like widgets or interactive
visualizations (ala %matplotlib notebook) to suddenly start working.

There is a lot more work to be done here, and possible it will never be possible
to have interactive widgets in the emacs notebook buffer, but I am hopeful that
one day we will at least be able to pop-up a browser window with a widget and
have it "talk" to the emacs notebook buffer.