ein-jupyter: Tweak deferred:wait factor when starting jupyter.
Not at all sure why, but succesfully starting jupyter and detecting login token
info is sensitive to this factor. On Windows, a call to accept-process-output
prior to the deferred chain also works, but unfortunately seems to cause issues
on some Linux installations.

For the moment a factor of 10 seems to work, but we will see what feedback from
the community is. This may need to be a customizable variable.