make test-compile and make install
test-compile attempts to egrep "bad" warnings with hardcoded strings.
This is better than nothing, but misses a lot.  We are close
to elimiinating this hack in favor of `(setq
byte-compile-error-on-warn t)` to enforce no warnings whatsoever.  Until
then, keep adding egrep strings to `test-compile`.

Without `(package-refresh-contents)`, the `make install` won't be able
to download and install dependencies for new users.  Remedy this.

Retire files that will never be used.