Fixes for ipdb support.
Somewhere along the line the local variables for ipdb buffers were getting
clobbered because they weren't permanent local, so now they are.

Also, somewhere along the way exit the debugger started generating just a status
message on the iopub channel, so we try to handle that more gracefully. It's
possible this will break with older versions of Jupyter - but backwards support
is apparently for weaklings in this modern age.