Initial support for control channel and kernel interrupt messages.
Control (no, not the video game) is a new messaging channel for communicating
with jupyter kernels. Introduced in version 5.3 of the messaging protocol, see
https://jupyter-client.readthedocs.io/en/stable/messaging.html#messages-on-the-control-router-dealer-channel
for more information.

It doesn't do much at the moment since, as far as I know, only the xeus-python
kernel actually responds to messages on this channel. As a test see the new
`ein:kernel-interrupt-command' function. With ipykernel nothing happens other
than to raise an error in the server log, but in xeus-python the kernel will
correctly be interrupted.

For more information on what the Jupyter foundation plans to do with this new
channel, see
https://blog.jupyter.org/a-visual-debugger-for-jupyter-914e61716559.