Basic misunderstanding about org-babel
As much of the code was copied from ob-ipython, we also copied its
mistakes.

Generally, one should not need to org-babel-remove-result before
org-babel-insert-result.  This was done by ob-ipython presumably to
avoid leaving the [....] "thinking indicator" behind when `:results
raw` was specified.

First, the thinking indicator is a visual decoration not in keeping
with org-babel convention.

Second, "raw" is raw, and we should expect duplicated results under that
directive.  We should not override "raw" with an opinionated
org-babel-remove result.

Finally, there was an egregious bug that showed only empty results
when ":results raw" was omitted (which should be the normal case!).

All in all, there was an ill-advised assumption about always using
":results raw drawer" simply because ob-ipython's example used it.

That assumption remains a crutch for all the ob-ein.feature tests.